fix: regress to one admin v1 file
