+++ basedosdados_api/api/v1/forms/__init__.py
@@ -1,3 +1,9 @@
 # -*- coding: utf-8 -*-
-from .reorder_columns_form import ReorderColumnsForm  # noqa: F401
-from .reorder_tables_form import ReorderTablesForm  # noqa: F401
+from basedosdados_api.api.v1.forms.admin_form import (  # noqa: F401
+    ColumnInlineForm,
+    CoverageInlineForm,
+    ObservationLevelForm,
+    TableInlineForm,
+)
+from basedosdados_api.api.v1.forms.reorder_columns_form import ReorderColumnsForm  # noqa: F401
+from basedosdados_api.api.v1.forms.reorder_tables_form import ReorderTablesForm  # noqa: F401
